building-ai-agents-with-gemini-dspy/pyproject.toml
[project]
name = "building-ai-agents-with-gemini-dspy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy>=3.0.1",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
]
